+++ client/src/actions/profileActions.js
@@ -1,7 +1,7 @@
 import { SESSION_ACTIONS } from "reducers/session";
 
-export const setOrganisationIndex = (index) => {
+export const setOrganisation = (id) => {
   return (dispatch) => {
-    dispatch({ type: SESSION_ACTIONS.SET_ORGANISATION_INDEX, payload: index });
+    dispatch({ type: SESSION_ACTIONS.SET_ORGANISATION_INDEX, payload: id });
   };
 };
