+++ client/src/reducers/session/selectors.js
@@ -1,9 +1,3 @@
-export const selectOrganisationId = (state) => {
-  const { user, organisationIndex } = state.session;
-  if (!user || isNaN(organisationIndex)) {
-    return null;
-  }
-  return user?.organisations[organisationIndex]?._id;
-};
+export const selectOrganisationId = (state) => state.session.organisationId;
 
 export const selectUser = (state) => state.session.user;
